refactor: rename index.tsx to PascalCase.tsx, add barrel index.tsx for S2-1c components
Matches the Tasklist convention established in S2-1a/b. CSS file renamed
from index.css to ResizablePanel.css and import updated accordingly.
